registration and login

+++ myapp/views/productViews.py
@@ -1,16 +1,23 @@
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, permission_classes
 from rest_framework.response import Response
 from rest_framework import status
 from myapp.Models import Product
 from myapp.serializers.productSerializer import ProductSerializer
+from myapp.permissions import IsAdminOrOwner
+from rest_framework.permissions import IsAuthenticated
 
 @api_view(['GET', 'POST'])
+@permission_classes([IsAuthenticated, IsAdminOrOwner])
 def product_list(request):
     if request.method == 'GET':
         products = Product.objects.all()
         serializer = ProductSerializer(products, many=True)
         return Response(serializer.data)
+    
     elif request.method == 'POST':
+        if not request.user.is_staff:
+            return Response({'detail': 'Not authorized to create products'}, status=status.HTTP_403_FORBIDDEN)
+        
         serializer = ProductSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -18,6 +25,7 @@ def product_list(request):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 @api_view(['GET', 'PUT', 'DELETE'])
+@permission_classes([IsAuthenticated, IsAdminOrOwner])
 def product_detail(request, pk):
     try:
         product = Product.objects.get(pk=pk)
@@ -27,12 +35,20 @@ def product_detail(request, pk):
     if request.method == 'GET':
         serializer = ProductSerializer(product)
         return Response(serializer.data)
+    
     elif request.method == 'PUT':
+        if not request.user.is_staff:
+            return Response({'detail': 'Not authorized to update products'}, status=status.HTTP_403_FORBIDDEN)
+        
         serializer = ProductSerializer(product, data=request.data)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    
     elif request.method == 'DELETE':
+        if not request.user.is_staff:
+            return Response({'detail': 'Not authorized to delete products'}, status=status.HTTP_403_FORBIDDEN)
+        
         product.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
